Fix package of moved GenericSettingProvider
