+++ facade/src/main/java/org/jboss/pnc/facade/impl/BuildTriggererImpl.java
@@ -45,7 +45,7 @@ import org.jboss.pnc.model.BuildConfiguration;
 import org.jboss.pnc.model.BuildConfigurationAudited;
 import org.jboss.pnc.model.BuildConfigurationSet;
 import org.jboss.pnc.model.IdRev;
-import org.jboss.pnc.rest.provider.GenericSettingProvider;
+import org.jboss.pnc.facade.providers.GenericSettingProvider;
 import org.jboss.pnc.spi.BuildOptions;
 import org.jboss.pnc.spi.coordinator.BuildCoordinator;
 import org.jboss.pnc.spi.coordinator.BuildSetTask;
